Have the runner skip JobActions that are not appropriate.
Use the OS metadata on a JobAction to determine if that action should be
run on a given machine based on whether it contains either 'any' or the OS
that the runner is running on.

Also allow explicit specification of the interpreter via the Interpreter
metadata on a JobAction for systems that do not have a built-in mechanism
for doing so.