Ditch using CFSSL for generating cert signing requests.
The bits we were using were totally self-contained anyways, so
I just copied in the functions and data structures we were using.
Code license for CFSSL is BSD 2-clause.

This saves a significant amount of space due to not pulling
in code we do not use anyways, and allows us to be Go 1.11 compatible
because cfssl has not made some Windows-related upstream changes for
Go 1.11 compatibility.