Fix lease invalidation bug.
Lease invalidation was resupting in lease objects that could not be
saved, leading to potential failures of dr-provision to restart when a
lease for a specific IP address was invalidated.  This fixes the lease
invalidation logic to handle invalid leases correctly, and fixes the
startup routine to handle the cases where a lease was invlaidated due
to IP address probing finding a conflict.